Refactor Libraries page to Pages folder
