+++ LibraryApplication/LibraryApplication/Pages/Libraries/Edit.aspx
@@ -1,4 +1,4 @@
-﻿<%@ Page Title="" Language="C#" MasterPageFile="~/Site.Master" AutoEventWireup="true" CodeBehind="EditLibrary.aspx.cs" Inherits="LibraryApplication.EditLibrary" %>
+﻿<%@ Page Title="" Language="C#" MasterPageFile="~/Site.Master" AutoEventWireup="true" CodeBehind="Edit.aspx.cs" Inherits="LibraryApplication.Pages.Libraries.Edit" %>
 
 <asp:Content ID="Content2" ContentPlaceHolderID="Main" runat="server">
     <div>
